+++ atropos/errors.py
@@ -26,6 +26,6 @@ class NotInAlphabetError(AtroposError):
     """
     Error that indicates an invalid character in an input sequence.
     """
-    def __init__(self, character: str):
-        super().__init__()
-        self.character = character
+    def __init__(self, characters: str):
+        super().__init__(f"Character(s) not in alphabet: {','.join(characters)}")
+        self.characters = characters